requirements.txt
progressbar2
tensorflow
keras
numpy